fix re-jit bug

+++ ssm/hmm/transitions.py
@@ -162,7 +162,7 @@ class SimpleStickyTransitions(Transitions):
         # If the transition matrix is given, recompute it given stay_probability.
         if transition_distribution is None:
             assert (stay_probability >= 0) and (stay_probability <= 1)
-            self.stay_probability = stay_probability
+            self.stay_probability = np.float32(stay_probability)  # prevent weak typing
             self._distribution = ssmd.Categorical(
                 logits=self._recompute_log_transition_matrix()
             )
